+++ packages/cater-memoization/plugin.js
@@ -1,0 +1,6 @@
+// Copyright Jon Williams 2017-2018. See LICENSE file.
+
+module.exports = function plugin() {
+  // Nothing to do yet, but in the future we'll need initialize backends like
+  // disk, memcache, etc.
+};